endpoint put regras

+++ projeto-ms/ms-calculate/src/main/java/org/yasmingv/ms_calculate/infra/RegrasRepositorio.java
@@ -3,5 +3,6 @@ package org.yasmingv.ms_calculate.infra;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.yasmingv.ms_calculate.dominio.Regras;
 
-public interface RegrasRepositorio extends JpaRepository<Regras, Integer> {
+public interface RegrasRepositorio extends JpaRepository<Regras, Long> {
+
 }
